renamed files, created first tests and initialze for Trips
