+++ ride_share.rb
@@ -2,6 +2,9 @@ require 'csv'
 
 module RideShare; end
 
-require_relative 'lib/drivers.rb'
-require_relative 'lib/trips.rb'
-require_relative 'lib/riders.rb'
+require_relative 'lib/driver.rb'
+require_relative 'lib/trip.rb'
+require_relative 'lib/rider.rb'
+
+class InvalidRatingError < StandardError
+end
