+++ specs/drivers_spec.rb
@@ -1,11 +1,11 @@
 require_relative 'spec_helper'
 
-describe 'tests' do
-  it "should run tests" do
-    true.must_equal false
-  end
-
-  it "should have access to drivers.rb" do
-    TEST_CONTSANT.must_equal 11
-  end
-end
+# describe 'tests' do
+#   it "should run tests" do
+#     true.must_equal false
+#   end
+#
+#   it "should have access to drivers.rb" do
+#     TEST_CONTSANT.must_equal 11
+#   end
+# end
